agregar types y arreglar estructura de directorio
