+++ src/renderer/src/types/Category.ts
@@ -1,0 +1,5 @@
+export interface Category {
+  id: string;
+  name: string;
+  description?: string;
+}
