+++ src/renderer/src/types/Product.ts
@@ -1,0 +1,9 @@
+export interface Product {
+  id: string;
+  name: string;
+  category: string;
+  stock: number;
+  price: number;
+  image?: string;
+  createdAt: Date;
+}
