+++ src/renderer/src/types/User.ts
@@ -1,0 +1,7 @@
+export interface User {
+  id: string;
+  name: string;
+  email: string;
+  role?: 'admin' | 'user';
+  createdAt: Date;
+}
